Memory managment working on qemu+opensbi

+++ src/sbi/timer.rs
@@ -14,6 +14,6 @@ pub fn reset_timer_trigger() {
     super::set_timer(get_time() + CLOCK_FREQ / TICKS_PER_SECOND);
 }
 
-pub fn get_time_ms() -> u64 {
-    return get_time() as u64 / (CLOCK_FREQ / MILLI_PER_SECOND) as u64;
+pub fn get_time_ms() -> f64 {
+    return get_time() as f64 / (CLOCK_FREQ / MILLI_PER_SECOND) as f64;
 }
